+++ src/main/java/step3/RacingGameController.java
@@ -1,0 +1,4 @@
+package step3;
+
+public class RacingGameController {
+}
