+++ src/main/java/step3/domain/RacingGame.java
@@ -1,4 +1,6 @@
-package step3;
+package step3.domain;
+
+import step3.view.ResultView;
 
 import java.util.ArrayList;
 import java.util.List;
